Fixed a few of the 'implicit conversion from 64-bit to 32-bit' warnings.  These were generally float values without the f so were being used as doubles.
Implemented the following TLMaths routines that simply wrap the standard maths versions and will eventually be platform specific:

Acosf( float )
Asinf( float ) 
Atanf( float )
Atan2f( float, float )

Modf( float )

Ceilf( float )
Floorf( float )

Also implemented the following type conversion routines that simply cast the type for now:

float 		S32ToFloat( s32 )
s32 		FloatToS32( float )

+++ Code/box2d/Source/Collision/Shapes/b2CircleShape.cpp
@@ -112,7 +112,7 @@ float32 b2CircleShape::ComputeSubmergedArea(	const b2Vec2& normal,
 	float32 r2 = m_radius*m_radius;
 	float32 l2 = l*l;
 	float32 area = r2 * (asinf(l/m_radius) + b2_pi/2.0f)+ l * b2Sqrt(r2 - l2);
-	float32 com = -2.0f/3.0f*pow(r2-l2,1.5f)/area;
+	float32 com = -2.0f/3.0f*powf(r2-l2,1.5f)/area;
 	
 	c->x = p.x + normal.x * com;
 	c->y = p.y + normal.y * com;
